Move submit buttons to separate component
