--- support-frontend/assets/pages/digital-subscription-checkout/components/SubscriptionSubmit.jsx
@@ -1,133 +1,0 @@
-// @flow
-
-// ----- Imports ----- //
-
-import type { Csrf as CsrfState } from 'helpers/csrf/csrfReducer';
-import React from 'react';
-import { connect } from 'react-redux';
-
-import { billingPeriodFromContrib, type ContributionType, getAmount, type PaymentMethod } from 'helpers/contributions';
-import { type IsoCurrency } from 'helpers/internationalisation/currency';
-import { type PaymentAuthorisation } from 'helpers/paymentIntegrations/readerRevenueApis';
-import type { SelectedAmounts } from 'helpers/contributions';
-import { getContributeButtonCopyWithPaymentType } from 'helpers/checkouts';
-import { hiddenIf } from 'helpers/utilities';
-import { setupRecurringPayPalPayment, type SetupPayPalRequestType } from 'helpers/paymentIntegrations/payPalRecurringCheckout';
-import type { BillingPeriod } from 'helpers/billingPeriods';
-import { PayPalExpressButton } from 'components/paypalExpressButton/PayPalExpressButton';
-import { type State } from '../digitalSubscriptionCheckoutReducer';
-import { sendFormSubmitEventForPayPalRecurring } from '../contributionsLandingActions';
-import { ButtonWithRightArrow } from './ButtonWithRightArrow/ButtonWithRightArrow';
-
-// ----- Types ----- //
-
-type PropTypes = {|
-  paymentMethod: PaymentMethod,
-  currency: IsoCurrency,
-  isWaiting: boolean,
-  currencyId: IsoCurrency,
-  csrf: CsrfState,
-  sendFormSubmitEventForPayPalRecurring: () => void,
-  setupRecurringPayPalPayment: SetupPayPalRequestType,
-  payPalHasLoaded: boolean,
-  isTestUser: boolean,
-  onPaymentAuthorisation: PaymentAuthorisation => void,
-  formIsSubmittable: boolean,
-  amount: number,
-  billingPeriod: BillingPeriod,
-|};
-
-function mapStateToProps(state: State) {
-  const { contributionType } = state.page.form;
-  return ({
-    currency: state.common.internationalisation.currencyId,
-    isWaiting: state.page.form.isWaiting,
-    paymentMethod: state.page.form.paymentMethod,
-    currencyId: state.common.internationalisation.currencyId,
-    csrf: state.page.csrf,
-    payPalHasLoaded: state.page.form.payPalHasLoaded,
-    isTestUser: state.page.user.isTestUser,
-    formIsSubmittable: state.page.form.formIsSubmittable,
-    amount: getAmount(
-      state.page.form.selectedAmounts,
-      state.page.form.formData.otherAmounts,
-      contributionType,
-    ),
-    billingPeriod: billingPeriodFromContrib(contributionType),
-  });
-}
-
-const mapDispatchToProps = (dispatch: Function) => ({
-  sendFormSubmitEventForPayPalRecurring: () => { dispatch(sendFormSubmitEventForPayPalRecurring()); },
-  setupRecurringPayPalPayment: (
-    resolve: Function,
-    reject: Function,
-    currencyId: IsoCurrency,
-    csrf: CsrfState,
-    amount: number,
-    billingPeriod: BillingPeriod,
-  ) => {
-    dispatch(setupRecurringPayPalPayment(resolve, reject, currencyId, csrf, amount, billingPeriod));
-  },
-});
-
-
-// ----- Render ----- //
-
-
-function SubscriptionSubmit(props: PropTypes) {
-
-  if (props.paymentMethod !== 'None') {
-    // if all payment methods are switched off, do not display the button
-    const formClassName = 'form--contribution';
-    const showPayPalRecurringButton = props.paymentMethod === 'PayPal' && props.contributionType !== 'ONE_OFF';
-
-    const submitButtonCopy = getContributeButtonCopyWithPaymentType(
-      props.contributionType,
-      props.otherAmount,
-      props.selectedAmounts,
-      props.currency,
-      props.paymentMethod,
-    );
-
-    // We have to show/hide PayPalExpressButton rather than conditionally rendering it
-    // because we don't want to destroy and replace the iframe each time.
-    // See PayPalExpressButtonsx for more info.
-    return (
-      <div className="form__submit">
-        <div
-          id="component-paypal-button-checkout"
-          className={hiddenIf(!showPayPalRecurringButton, 'component-paypal-button-checkout')}
-        >
-          <PayPalExpressButton
-            onPaymentAuthorisation={props.onPaymentAuthorisation}
-            csrf={props.csrf}
-            currencyId={props.currencyId}
-            hasLoaded={props.payPalHasLoaded}
-            canOpen={() => props.formIsSubmittable}
-            onClick={() => props.sendFormSubmitEventForPayPalRecurring()}
-            formClassName={formClassName}
-            isTestUser={props.isTestUser}
-            setupRecurringPayPalPayment={props.setupRecurringPayPalPayment}
-            amount={props.amount}
-            billingPeriod={props.billingPeriod}
-          />
-        </div>
-        <ButtonWithRightArrow
-          accessibilityHintId="accessibility-hint-submit-contribution"
-          disabled={props.isWaiting}
-          type="submit"
-          componentClassName="form__submit--contribution"
-          buttonClassName={hiddenIf(showPayPalRecurringButton, 'form__submit-button')}
-          buttonCopy={submitButtonCopy}
-        />
-      </div>
-    );
-  }
-
-  return null;
-}
-
-const NewContributionSubmit = connect(mapStateToProps, mapDispatchToProps)(SubscriptionSubmit);
-
-export { NewContributionSubmit };
